Revert commits
The recent PR were suppose to be parts of more PR to get the entire codebase complete, or at least
where i want it to be and in a better working state for easier update and maintenance.
Ive come to the point that it would take a lot of time and effort
to get all PRs done since each one has to be modified to suit the current codebase, causing
a few minor but suppose to be fixable conversion errors. The repo have diverted too much
from the time i made the original fork, causing each merges conflicts and needing manual intervention.
The recent incident, the reluctance to accept the changes, the consideration of issue without details,
the need for third-party confirmation didnt help in giving motivation to do so and complete this.
I have worked a long time not just this repo, but on other cores fixing shit nobody else wanted to take
and it still amazes me that my intention are getting questioned in favour of.....
I would just rather continue working on my fork rather than starting over.
Reverting this would be more beneficial for everyone.
History is kept(i hoped), for reference.